fix: add assertions to tests
